+++ src/main/java/de/dm/intellij/liferay/language/freemarker/LiferayApplicationDisplayTemplateLiveTemplatesProvider.java
@@ -1,0 +1,21 @@
+package de.dm.intellij.liferay.language.freemarker;
+
+import com.intellij.codeInsight.template.impl.DefaultLiveTemplatesProvider;
+import org.jetbrains.annotations.Nullable;
+
+/**
+ * Adds Live Templates for Application Display Templates
+ */
+public class LiferayApplicationDisplayTemplateLiveTemplatesProvider implements DefaultLiveTemplatesProvider {
+
+    @Override
+    public String[] getDefaultLiveTemplateFiles() {
+        return new String[]{"liveTemplates/ApplicationDisplayTemplate"};
+    }
+
+    @Nullable
+    @Override
+    public String[] getHiddenLiveTemplateFiles() {
+        return new String[0];
+    }
+}